chp6_adivinhacao.py
This code refers to the game at the end of chapter 6 of this course. 

It was changed the variable 'numero_secreto' value was to assume a random number instead of a fixed number defined for the programmer. This class has turned the game more interesting. Further, has taught how to import and use the module "random" of Python.