chp7_adivinhacao.py - end of the game
This is the last version of the 'guessing' game presented at this Alura course. On This class was added different levels for the game and was also implemented a points system based on the distance between the trial value and the random value. The function abs( ) was introduced to calculate the variable 'lost_points' that represents the lost points in the trial.