Update hangman.py and Create words.txt
Working with files in Python. The file 'words.txt' contain names of fruits, that will be used in the game, this file was created with the following commands:

file = open("words.txt", "w")

and to write on file

file.write("fruit_name\n")

The code in hangman.py was updated to use a random word got from the .txt file, using the random library. The function strip( ) was used to remove the character '\n' for the selected word.