Update account.py - Static Method
Here was created a static method bank_code( ), that returns codes of differents Brazilian banks. These codes are statics, i.e., doesn't change after object instantiate, will be always the same. Therefore, a static method can be called without an object instantiating. It's noteworthy that this is the last commit of this course.